+++ src/common/dto/iam.dto.ts
@@ -1,3 +1,0 @@
-export class IamDto {
-  iam: string
-}