+++ src/core/core.module.ts
@@ -1,7 +1,0 @@
-import { Module } from "@nestjs/common"
-import { CoreController } from "./core.controller"
-
-@Module({
-  controllers: [CoreController],
-})
-export class CoreModule {}