+++ src/renderer/renderer.module.ts
@@ -1,10 +1,11 @@
 import { Module } from "@nestjs/common"
+import { BlocklistModule } from "../blocklist/blocklist.module"
 import { HeadlessBrowserModule } from "../headless-browser/headless-browser.module"
 import { RendererController } from "./renderer.controller"
 import { RendererService } from "./renderer.service"
 
 @Module({
-  imports: [HeadlessBrowserModule],
+  imports: [BlocklistModule, HeadlessBrowserModule],
   controllers: [RendererController],
   providers: [RendererService],
 })